refactor environment auth service layout
