+++ apps/desktop/src/ipc/methods/sshEnvironment.ts
@@ -1,4 +1,10 @@
 import {
+  bootstrapRemoteBearerSession,
+  fetchRemoteEnvironmentDescriptor,
+  fetchRemoteSessionState,
+  issueRemoteWebSocketTicket,
+} from "@t3tools/client-runtime";
+import {
   DesktopDiscoveredSshHostSchema,
   DesktopSshBearerBootstrapInputSchema,
   DesktopSshBearerRequestInputSchema,
@@ -13,6 +19,8 @@ import {
   AuthSessionState,
   AuthWebSocketTicketResult,
 } from "@t3tools/contracts";
+import { resolveLoopbackSshHttpBaseUrl } from "@t3tools/ssh/tunnel";
+import * as Data from "effect/Data";
 import * as Effect from "effect/Effect";
 import * as Schema from "effect/Schema";
 
@@ -20,7 +28,34 @@ import * as IpcChannels from "../channels.ts";
 import { makeIpcMethod } from "../DesktopIpc.ts";
 import * as DesktopSshEnvironment from "../../ssh/DesktopSshEnvironment.ts";
 import * as DesktopSshPasswordPrompts from "../../ssh/DesktopSshPasswordPrompts.ts";
-import * as DesktopSshRemoteApi from "../../ssh/DesktopSshRemoteApi.ts";
+
+type DesktopSshEnvironmentRequestOperation =
+  | "fetch-environment-descriptor"
+  | "bootstrap-bearer-session"
+  | "fetch-session-state"
+  | "issue-websocket-ticket";
+
+export class DesktopSshEnvironmentRequestError extends Data.TaggedError(
+  "DesktopSshEnvironmentRequestError",
+)<{
+  readonly operation: DesktopSshEnvironmentRequestOperation;
+  readonly cause: unknown;
+}> {
+  override get message() {
+    return `SSH remote API request failed during ${this.operation}.`;
+  }
+}
+
+const withLoopbackSshApi =
+  <A, E, R>(
+    operation: DesktopSshEnvironmentRequestOperation,
+    use: (httpBaseUrl: string) => Effect.Effect<A, E, R>,
+  ) =>
+  (httpBaseUrl: string): Effect.Effect<A, DesktopSshEnvironmentRequestError, R> =>
+    resolveLoopbackSshHttpBaseUrl(httpBaseUrl).pipe(
+      Effect.flatMap(use),
+      Effect.mapError((cause) => new DesktopSshEnvironmentRequestError({ operation, cause })),
+    );
 
 export const discoverSshHosts = makeIpcMethod({
   channel: IpcChannels.DISCOVER_SSH_HOSTS_CHANNEL,
@@ -69,8 +104,9 @@ export const fetchSshEnvironmentDescriptor = makeIpcMethod({
   payload: DesktopSshHttpBaseUrlInputSchema,
   result: ExecutionEnvironmentDescriptor,
   handler: Effect.fn("desktop.ipc.sshEnvironment.fetchDescriptor")(function* ({ httpBaseUrl }) {
-    const remoteApi = yield* DesktopSshRemoteApi.DesktopSshRemoteApi;
-    return yield* remoteApi.fetchEnvironmentDescriptor({ httpBaseUrl });
+    return yield* withLoopbackSshApi("fetch-environment-descriptor", (resolvedHttpBaseUrl) =>
+      fetchRemoteEnvironmentDescriptor({ httpBaseUrl: resolvedHttpBaseUrl }),
+    )(httpBaseUrl);
   }),
 });
 
@@ -82,8 +118,12 @@ export const bootstrapSshBearerSession = makeIpcMethod({
     httpBaseUrl,
     credential,
   }) {
-    const remoteApi = yield* DesktopSshRemoteApi.DesktopSshRemoteApi;
-    return yield* remoteApi.bootstrapBearerSession({ httpBaseUrl, credential });
+    return yield* withLoopbackSshApi("bootstrap-bearer-session", (resolvedHttpBaseUrl) =>
+      bootstrapRemoteBearerSession({
+        httpBaseUrl: resolvedHttpBaseUrl,
+        credential,
+      }),
+    )(httpBaseUrl);
   }),
 });
 
@@ -95,8 +135,12 @@ export const fetchSshSessionState = makeIpcMethod({
     httpBaseUrl,
     bearerToken,
   }) {
-    const remoteApi = yield* DesktopSshRemoteApi.DesktopSshRemoteApi;
-    return yield* remoteApi.fetchSessionState({ httpBaseUrl, bearerToken });
+    return yield* withLoopbackSshApi("fetch-session-state", (resolvedHttpBaseUrl) =>
+      fetchRemoteSessionState({
+        httpBaseUrl: resolvedHttpBaseUrl,
+        bearerToken,
+      }),
+    )(httpBaseUrl);
   }),
 });
 
@@ -108,8 +152,12 @@ export const issueSshWebSocketTicket = makeIpcMethod({
     httpBaseUrl,
     bearerToken,
   }) {
-    const remoteApi = yield* DesktopSshRemoteApi.DesktopSshRemoteApi;
-    return yield* remoteApi.issueWebSocketTicket({ httpBaseUrl, bearerToken });
+    return yield* withLoopbackSshApi("issue-websocket-ticket", (resolvedHttpBaseUrl) =>
+      issueRemoteWebSocketTicket({
+        httpBaseUrl: resolvedHttpBaseUrl,
+        bearerToken,
+      }),
+    )(httpBaseUrl);
   }),
 });
