+++ apps/server/src/auth/EnvironmentAuthPolicy.test.ts
@@ -3,13 +3,12 @@ import { expect, it } from "@effect/vitest";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 
-import type { ServerConfigShape } from "../../config.ts";
-import { ServerConfig } from "../../config.ts";
-import { ServerAuthPolicy } from "../Services/ServerAuthPolicy.ts";
-import { ServerAuthPolicyLive } from "./ServerAuthPolicy.ts";
+import type { ServerConfigShape } from "../config.ts";
+import { ServerConfig } from "../config.ts";
+import * as EnvironmentAuthPolicy from "./EnvironmentAuthPolicy.ts";
 
-const makeServerAuthPolicyLayer = (overrides?: Partial<ServerConfigShape>) =>
-  ServerAuthPolicyLive.pipe(
+const makeEnvironmentAuthPolicyLayer = (overrides?: Partial<ServerConfigShape>) =>
+  EnvironmentAuthPolicy.layer.pipe(
     Layer.provide(
       Layer.effect(
         ServerConfig,
@@ -26,10 +25,10 @@ const makeServerAuthPolicyLayer = (overrides?: Partial<ServerConfigShape>) =>
     ),
   );
 
-it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
+it.layer(NodeServices.layer)("EnvironmentAuthPolicy.layer", (it) => {
   it.effect("uses desktop-managed-local policy for desktop mode", () =>
     Effect.gen(function* () {
-      const policy = yield* ServerAuthPolicy;
+      const policy = yield* EnvironmentAuthPolicy.EnvironmentAuthPolicy;
       const descriptor = yield* policy.getDescriptor();
 
       expect(descriptor.policy).toBe("desktop-managed-local");
@@ -37,7 +36,7 @@ it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
       expect(descriptor.sessionCookieName).toBe("t3_session_3773");
     }).pipe(
       Effect.provide(
-        makeServerAuthPolicyLayer({
+        makeEnvironmentAuthPolicyLayer({
           mode: "desktop",
           port: 3773,
         }),
@@ -47,14 +46,14 @@ it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
 
   it.effect("uses remote-reachable policy for desktop mode when bound beyond loopback", () =>
     Effect.gen(function* () {
-      const policy = yield* ServerAuthPolicy;
+      const policy = yield* EnvironmentAuthPolicy.EnvironmentAuthPolicy;
       const descriptor = yield* policy.getDescriptor();
 
       expect(descriptor.policy).toBe("remote-reachable");
       expect(descriptor.bootstrapMethods).toEqual(["desktop-bootstrap", "one-time-token"]);
     }).pipe(
       Effect.provide(
-        makeServerAuthPolicyLayer({
+        makeEnvironmentAuthPolicyLayer({
           mode: "desktop",
           host: "0.0.0.0",
         }),
@@ -64,7 +63,7 @@ it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
 
   it.effect("uses loopback-browser policy for loopback web hosts", () =>
     Effect.gen(function* () {
-      const policy = yield* ServerAuthPolicy;
+      const policy = yield* EnvironmentAuthPolicy.EnvironmentAuthPolicy;
       const descriptor = yield* policy.getDescriptor();
 
       expect(descriptor.policy).toBe("loopback-browser");
@@ -72,7 +71,7 @@ it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
       expect(descriptor.sessionCookieName).toBe("t3_session");
     }).pipe(
       Effect.provide(
-        makeServerAuthPolicyLayer({
+        makeEnvironmentAuthPolicyLayer({
           mode: "web",
           host: "127.0.0.1",
         }),
@@ -82,14 +81,14 @@ it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
 
   it.effect("uses remote-reachable policy for wildcard web hosts", () =>
     Effect.gen(function* () {
-      const policy = yield* ServerAuthPolicy;
+      const policy = yield* EnvironmentAuthPolicy.EnvironmentAuthPolicy;
       const descriptor = yield* policy.getDescriptor();
 
       expect(descriptor.policy).toBe("remote-reachable");
       expect(descriptor.bootstrapMethods).toEqual(["one-time-token"]);
     }).pipe(
       Effect.provide(
-        makeServerAuthPolicyLayer({
+        makeEnvironmentAuthPolicyLayer({
           mode: "web",
           host: "0.0.0.0",
         }),
@@ -99,13 +98,13 @@ it.layer(NodeServices.layer)("ServerAuthPolicyLive", (it) => {
 
   it.effect("uses remote-reachable policy for non-loopback web hosts", () =>
     Effect.gen(function* () {
-      const policy = yield* ServerAuthPolicy;
+      const policy = yield* EnvironmentAuthPolicy.EnvironmentAuthPolicy;
       const descriptor = yield* policy.getDescriptor();
 
       expect(descriptor.policy).toBe("remote-reachable");
     }).pipe(
       Effect.provide(
-        makeServerAuthPolicyLayer({
+        makeEnvironmentAuthPolicyLayer({
           mode: "web",
           host: "192.168.1.50",
         }),
