+++ apps/server/src/auth/EnvironmentAuthPolicy.ts
@@ -1,13 +1,22 @@
 import type { ServerAuthDescriptor } from "@t3tools/contracts";
+import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 
-import { ServerConfig } from "../../config.ts";
-import { ServerAuthPolicy, type ServerAuthPolicyShape } from "../Services/ServerAuthPolicy.ts";
-import { resolveSessionCookieName } from "../utils.ts";
-import { isLoopbackHost, isWildcardHost } from "../../startupAccess.ts";
+import { ServerConfig } from "../config.ts";
+import { resolveSessionCookieName } from "./utils.ts";
+import { isLoopbackHost, isWildcardHost } from "../startupAccess.ts";
 
-export const makeServerAuthPolicy = Effect.gen(function* () {
+export interface EnvironmentAuthPolicyShape {
+  readonly getDescriptor: () => Effect.Effect<ServerAuthDescriptor>;
+}
+
+export class EnvironmentAuthPolicy extends Context.Service<
+  EnvironmentAuthPolicy,
+  EnvironmentAuthPolicyShape
+>()("t3/auth/EnvironmentAuthPolicy") {}
+
+export const make = Effect.fn("makeEnvironmentAuthPolicy")(function* () {
   const config = yield* ServerConfig;
   const isRemoteReachable = isWildcardHost(config.host) || !isLoopbackHost(config.host);
 
@@ -39,7 +48,7 @@ export const makeServerAuthPolicy = Effect.gen(function* () {
 
   return {
     getDescriptor: () => Effect.succeed(descriptor),
-  } satisfies ServerAuthPolicyShape;
+  } satisfies EnvironmentAuthPolicyShape;
 });
 
-export const ServerAuthPolicyLive = Layer.effect(ServerAuthPolicy, makeServerAuthPolicy);
+export const layer = Layer.effect(EnvironmentAuthPolicy, make());
