--- apps/server/src/auth/Layers/ServerAuth.ts
@@ -1,445 +1,0 @@
-import {
-  AuthAccessTokenType,
-  AuthAdministrativeScopes,
-  AuthStandardClientScopes,
-  type AuthAccessTokenResult,
-  type AuthClientSession,
-  type AuthBrowserSessionResult,
-  type AuthEnvironmentScope,
-  type AuthPairingCredentialResult,
-  type AuthSessionState,
-  type AuthWebSocketTicketResult,
-} from "@t3tools/contracts";
-import { encodeOAuthScope } from "@t3tools/shared/oauthScope";
-import * as DateTime from "effect/DateTime";
-import * as Effect from "effect/Effect";
-import * as Layer from "effect/Layer";
-import * as Option from "effect/Option";
-import * as HttpServerRequest from "effect/unstable/http/HttpServerRequest";
-
-import {
-  AuthControlPlane,
-  INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT,
-} from "../Services/AuthControlPlane.ts";
-import { ServerAuthPolicyLive } from "./ServerAuthPolicy.ts";
-import {
-  type BootstrapCredentialError,
-  BootstrapCredentialService,
-} from "../Services/BootstrapCredentialService.ts";
-import { ServerAuthPolicy } from "../Services/ServerAuthPolicy.ts";
-import {
-  ServerAuth,
-  type AuthenticatedSession,
-  ServerAuthForbiddenOperationError,
-  ServerAuthInternalError,
-  ServerAuthInvalidCredentialError,
-  ServerAuthInvalidRequestError,
-  type ServerAuthShape,
-} from "../Services/ServerAuth.ts";
-import {
-  type SessionCredentialError,
-  SessionCredentialService,
-} from "../Services/SessionCredentialService.ts";
-import { AuthControlPlaneLive, AuthCoreLive } from "./AuthControlPlane.ts";
-
-type BootstrapExchangeResult = {
-  readonly response: AuthBrowserSessionResult;
-  readonly sessionToken: string;
-};
-
-const AUTHORIZATION_PREFIX = "Bearer ";
-const WEBSOCKET_TICKET_QUERY_PARAM = "wsTicket";
-
-export function toBootstrapExchangeError(
-  cause: BootstrapCredentialError,
-): ServerAuthInvalidCredentialError | ServerAuthInternalError {
-  if (cause._tag === "BootstrapCredentialInternalError") {
-    return new ServerAuthInternalError({
-      message: "Failed to validate bootstrap credential.",
-      cause,
-    });
-  }
-
-  return new ServerAuthInvalidCredentialError({
-    reason: "invalid_credential",
-    cause,
-  });
-}
-
-const mapSessionVerificationErrors = <A, R>(
-  effect: Effect.Effect<A, SessionCredentialError, R>,
-): Effect.Effect<A, ServerAuthInvalidCredentialError | ServerAuthInternalError, R> =>
-  effect.pipe(
-    Effect.catchTags({
-      SessionCredentialInvalidError: (cause) =>
-        Effect.fail(new ServerAuthInvalidCredentialError({ reason: "invalid_credential", cause })),
-      SessionCredentialInternalError: (cause) =>
-        Effect.fail(
-          new ServerAuthInternalError({
-            message: "Failed to validate session credential.",
-            cause,
-          }),
-        ),
-    }),
-  );
-
-function parseBearerToken(request: HttpServerRequest.HttpServerRequest): string | null {
-  const header = request.headers["authorization"];
-  if (typeof header !== "string" || !header.startsWith(AUTHORIZATION_PREFIX)) {
-    return null;
-  }
-  const token = header.slice(AUTHORIZATION_PREFIX.length).trim();
-  return token.length > 0 ? token : null;
-}
-
-export const makeServerAuth = Effect.gen(function* () {
-  const policy = yield* ServerAuthPolicy;
-  const bootstrapCredentials = yield* BootstrapCredentialService;
-  const authControlPlane = yield* AuthControlPlane;
-  const sessions = yield* SessionCredentialService;
-  const descriptor = yield* policy.getDescriptor();
-
-  const authenticateToken = (
-    token: string,
-  ): Effect.Effect<
-    AuthenticatedSession,
-    ServerAuthInvalidCredentialError | ServerAuthInternalError
-  > =>
-    sessions.verify(token).pipe(
-      Effect.tapErrorTag("SessionCredentialInvalidError", (cause) =>
-        Effect.logWarning("Rejected authenticated session credential.").pipe(
-          Effect.annotateLogs({
-            reason: cause.message,
-          }),
-        ),
-      ),
-      Effect.map((session) => ({
-        sessionId: session.sessionId,
-        subject: session.subject,
-        method: session.method,
-        scopes: session.scopes,
-        ...(session.expiresAt ? { expiresAt: session.expiresAt } : {}),
-      })),
-      mapSessionVerificationErrors,
-    );
-
-  const authenticateRequest = (request: HttpServerRequest.HttpServerRequest) => {
-    const cookieToken = request.cookies[sessions.cookieName];
-    const bearerToken = parseBearerToken(request);
-    const credential = cookieToken ?? bearerToken;
-    if (!credential) {
-      return Effect.fail(new ServerAuthInvalidCredentialError({ reason: "missing_credential" }));
-    }
-    return authenticateToken(credential);
-  };
-
-  const getSessionState: ServerAuthShape["getSessionState"] = (request) =>
-    authenticateRequest(request).pipe(
-      Effect.map(
-        (session) =>
-          ({
-            authenticated: true,
-            auth: descriptor,
-            scopes: session.scopes,
-            sessionMethod: session.method,
-            ...(session.expiresAt ? { expiresAt: DateTime.toUtc(session.expiresAt) } : {}),
-          }) satisfies AuthSessionState,
-      ),
-      Effect.catchTag("ServerAuthInvalidCredentialError", () =>
-        Effect.succeed({
-          authenticated: false,
-          auth: descriptor,
-        } satisfies AuthSessionState),
-      ),
-    );
-
-  const createBrowserSession: ServerAuthShape["createBrowserSession"] = (
-    credential,
-    requestMetadata,
-  ) =>
-    bootstrapCredentials.consume(credential).pipe(
-      Effect.mapError(toBootstrapExchangeError),
-      Effect.flatMap((grant) =>
-        sessions
-          .issue({
-            method: "browser-session-cookie",
-            subject: grant.subject,
-            scopes: grant.scopes,
-            client: {
-              ...requestMetadata,
-              ...(grant.label ? { label: grant.label } : {}),
-            },
-          })
-          .pipe(
-            Effect.mapError(
-              (cause) =>
-                new ServerAuthInternalError({
-                  message: "Failed to issue authenticated session.",
-                  cause,
-                }),
-            ),
-          ),
-      ),
-      Effect.map(
-        (session) =>
-          ({
-            response: {
-              authenticated: true,
-              scopes: session.scopes,
-              sessionMethod: session.method,
-              expiresAt: DateTime.toUtc(session.expiresAt),
-            } satisfies AuthBrowserSessionResult,
-            sessionToken: session.token,
-          }) satisfies BootstrapExchangeResult,
-      ),
-    );
-
-  const exchangeBootstrapCredentialForAccessToken: ServerAuthShape["exchangeBootstrapCredentialForAccessToken"] =
-    (credential, requestedScopes, requestMetadata) =>
-      bootstrapCredentials.consume(credential).pipe(
-        Effect.mapError(toBootstrapExchangeError),
-        Effect.flatMap((grant) =>
-          Effect.gen(function* () {
-            const grantedScopes = requestedScopes ?? grant.scopes;
-            if (!grantedScopes.every((scope) => grant.scopes.includes(scope))) {
-              return yield* new ServerAuthInvalidRequestError({
-                reason: "scope_not_granted",
-              });
-            }
-            return yield* sessions
-              .issue({
-                method: "bearer-access-token",
-                subject: grant.subject,
-                scopes: grantedScopes,
-                client: {
-                  ...requestMetadata,
-                  ...(grant.label ? { label: grant.label } : {}),
-                },
-              })
-              .pipe(
-                Effect.mapError(
-                  (cause) =>
-                    new ServerAuthInternalError({
-                      message: "Failed to issue authenticated access token.",
-                      cause,
-                    }),
-                ),
-              );
-          }),
-        ),
-        Effect.flatMap((session) =>
-          DateTime.now.pipe(
-            Effect.map(
-              (now) =>
-                ({
-                  access_token: session.token,
-                  issued_token_type: AuthAccessTokenType,
-                  token_type: "Bearer",
-                  expires_in: Math.max(
-                    0,
-                    Math.floor(
-                      (session.expiresAt.epochMilliseconds - now.epochMilliseconds) / 1000,
-                    ),
-                  ),
-                  scope: encodeOAuthScope(session.scopes),
-                }) satisfies AuthAccessTokenResult,
-            ),
-          ),
-        ),
-      );
-
-  const issuePairingCredentialForSubject = (input: {
-    readonly scopes: ReadonlyArray<AuthEnvironmentScope>;
-    readonly subject: string;
-    readonly label?: string;
-  }) =>
-    authControlPlane
-      .createPairingLink({
-        scopes: input.scopes,
-        subject: input.subject,
-        ...(input.label ? { label: input.label } : {}),
-      })
-      .pipe(
-        Effect.mapError(
-          (cause) =>
-            new ServerAuthInternalError({
-              message: "Failed to issue pairing credential.",
-              cause,
-            }),
-        ),
-        Effect.map(
-          (issued) =>
-            ({
-              id: issued.id,
-              credential: issued.credential,
-              ...(issued.label ? { label: issued.label } : {}),
-              expiresAt: issued.expiresAt,
-            }) satisfies AuthPairingCredentialResult,
-        ),
-      );
-
-  const issuePairingCredential: ServerAuthShape["issuePairingCredential"] = (input) =>
-    issuePairingCredentialForSubject({
-      scopes: input?.scopes ?? AuthStandardClientScopes,
-      subject: "one-time-token",
-      ...(input?.label ? { label: input.label } : {}),
-    });
-
-  const issueStartupPairingCredential: ServerAuthShape["issueStartupPairingCredential"] = () =>
-    issuePairingCredentialForSubject({
-      scopes: AuthAdministrativeScopes,
-      subject: INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT,
-    });
-
-  const listPairingLinks: ServerAuthShape["listPairingLinks"] = () =>
-    authControlPlane
-      .listPairingLinks({ excludeSubjects: [INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT] })
-      .pipe(
-        Effect.mapError(
-          (cause) =>
-            new ServerAuthInternalError({
-              message: "Failed to load pairing links.",
-              cause,
-            }),
-        ),
-      );
-
-  const revokePairingLink: ServerAuthShape["revokePairingLink"] = (id) =>
-    authControlPlane.revokePairingLink(id).pipe(
-      Effect.mapError(
-        (cause) =>
-          new ServerAuthInternalError({
-            message: "Failed to revoke pairing link.",
-            cause,
-          }),
-      ),
-    );
-
-  const listClientSessions: ServerAuthShape["listClientSessions"] = (currentSessionId) =>
-    authControlPlane.listSessions().pipe(
-      Effect.mapError(
-        (cause) =>
-          new ServerAuthInternalError({
-            message: "Failed to load paired clients.",
-            cause,
-          }),
-      ),
-      Effect.map((clientSessions) =>
-        clientSessions.map(
-          (clientSession): AuthClientSession => ({
-            ...clientSession,
-            current: clientSession.sessionId === currentSessionId,
-          }),
-        ),
-      ),
-    );
-
-  const revokeClientSession: ServerAuthShape["revokeClientSession"] = (
-    currentSessionId,
-    targetSessionId,
-  ) =>
-    Effect.gen(function* () {
-      if (currentSessionId === targetSessionId) {
-        return yield* new ServerAuthForbiddenOperationError({
-          reason: "current_session_revoke_not_allowed",
-        });
-      }
-      return yield* authControlPlane.revokeSession(targetSessionId).pipe(
-        Effect.mapError(
-          (cause) =>
-            new ServerAuthInternalError({
-              message: "Failed to revoke client session.",
-              cause,
-            }),
-        ),
-      );
-    });
-
-  const revokeOtherClientSessions: ServerAuthShape["revokeOtherClientSessions"] = (
-    currentSessionId,
-  ) =>
-    authControlPlane.revokeOtherSessionsExcept(currentSessionId).pipe(
-      Effect.mapError(
-        (cause) =>
-          new ServerAuthInternalError({
-            message: "Failed to revoke other client sessions.",
-            cause,
-          }),
-      ),
-    );
-
-  const issueStartupPairingUrl: ServerAuthShape["issueStartupPairingUrl"] = (baseUrl) =>
-    issueStartupPairingCredential().pipe(
-      Effect.map((issued) => {
-        const url = new URL(baseUrl);
-        url.pathname = "/pair";
-        url.searchParams.delete("token");
-        url.hash = new URLSearchParams([["token", issued.credential]]).toString();
-        return url.toString();
-      }),
-    );
-
-  const issueWebSocketTicket: ServerAuthShape["issueWebSocketTicket"] = (session) =>
-    sessions.issueWebSocketToken(session.sessionId).pipe(
-      Effect.mapError(
-        (cause) =>
-          new ServerAuthInternalError({
-            message: "Failed to issue websocket token.",
-            cause,
-          }),
-      ),
-      Effect.map(
-        (issued) =>
-          ({
-            ticket: issued.token,
-            expiresAt: DateTime.toUtc(issued.expiresAt),
-          }) satisfies AuthWebSocketTicketResult,
-      ),
-    );
-
-  const authenticateWebSocketUpgrade: ServerAuthShape["authenticateWebSocketUpgrade"] = (request) =>
-    Effect.gen(function* () {
-      const requestUrl = HttpServerRequest.toURL(request);
-      if (Option.isSome(requestUrl)) {
-        const websocketTicket = requestUrl.value.searchParams.get(WEBSOCKET_TICKET_QUERY_PARAM);
-        if (websocketTicket && websocketTicket.trim().length > 0) {
-          return yield* sessions.verifyWebSocketToken(websocketTicket).pipe(
-            Effect.map((session) => ({
-              sessionId: session.sessionId,
-              subject: session.subject,
-              method: session.method,
-              scopes: session.scopes,
-              ...(session.expiresAt ? { expiresAt: session.expiresAt } : {}),
-            })),
-            mapSessionVerificationErrors,
-          );
-        }
-      }
-
-      return yield* authenticateRequest(request);
-    });
-
-  return {
-    getDescriptor: () => Effect.succeed(descriptor),
-    getSessionState,
-    createBrowserSession,
-    exchangeBootstrapCredentialForAccessToken,
-    issuePairingCredential,
-    issueStartupPairingCredential,
-    listPairingLinks,
-    revokePairingLink,
-    listClientSessions,
-    revokeClientSession,
-    revokeOtherClientSessions,
-    authenticateHttpRequest: authenticateRequest,
-    authenticateWebSocketUpgrade,
-    issueWebSocketTicket,
-    issueStartupPairingUrl,
-  } satisfies ServerAuthShape;
-});
-
-export const ServerAuthLive = Layer.effect(ServerAuth, makeServerAuth).pipe(
-  Layer.provideMerge(AuthControlPlaneLive),
-  Layer.provideMerge(AuthCoreLive),
-  Layer.provideMerge(ServerAuthPolicyLive),
-);
