--- apps/server/src/auth/Services/AuthControlPlane.ts
@@ -1,73 +1,0 @@
-import type {
-  AuthClientMetadata,
-  AuthClientSession,
-  AuthEnvironmentScope,
-  AuthPairingLink,
-  AuthSessionId,
-} from "@t3tools/contracts";
-import * as Data from "effect/Data";
-import * as DateTime from "effect/DateTime";
-import * as Duration from "effect/Duration";
-import * as Effect from "effect/Effect";
-import * as Context from "effect/Context";
-
-export const DEFAULT_SESSION_SUBJECT = "cli-issued-session";
-export const INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT = "administrative-bootstrap";
-
-export interface IssuedPairingLink {
-  readonly id: string;
-  readonly credential: string;
-  readonly scopes: ReadonlyArray<AuthEnvironmentScope>;
-  readonly subject: string;
-  readonly label?: string;
-  readonly createdAt: DateTime.Utc;
-  readonly expiresAt: DateTime.Utc;
-}
-
-export interface IssuedBearerSession {
-  readonly sessionId: AuthSessionId;
-  readonly token: string;
-  readonly method: "bearer-access-token";
-  readonly scopes: ReadonlyArray<AuthEnvironmentScope>;
-  readonly subject: string;
-  readonly client: AuthClientMetadata;
-  readonly expiresAt: DateTime.Utc;
-}
-
-export class AuthControlPlaneError extends Data.TaggedError("AuthControlPlaneError")<{
-  readonly message: string;
-  readonly cause?: unknown;
-}> {}
-
-export interface AuthControlPlaneShape {
-  readonly createPairingLink: (input?: {
-    readonly ttl?: Duration.Duration;
-    readonly label?: string;
-    readonly scopes?: ReadonlyArray<AuthEnvironmentScope>;
-    readonly subject?: string;
-  }) => Effect.Effect<IssuedPairingLink, AuthControlPlaneError>;
-  readonly listPairingLinks: (input?: {
-    readonly excludeSubjects?: ReadonlyArray<string>;
-  }) => Effect.Effect<ReadonlyArray<AuthPairingLink>, AuthControlPlaneError>;
-  readonly revokePairingLink: (id: string) => Effect.Effect<boolean, AuthControlPlaneError>;
-  readonly issueSession: (input?: {
-    readonly ttl?: Duration.Duration;
-    readonly subject?: string;
-    readonly scopes?: ReadonlyArray<AuthEnvironmentScope>;
-    readonly label?: string;
-  }) => Effect.Effect<IssuedBearerSession, AuthControlPlaneError>;
-  readonly listSessions: () => Effect.Effect<
-    ReadonlyArray<AuthClientSession>,
-    AuthControlPlaneError
-  >;
-  readonly revokeSession: (
-    sessionId: AuthSessionId,
-  ) => Effect.Effect<boolean, AuthControlPlaneError>;
-  readonly revokeOtherSessionsExcept: (
-    sessionId: AuthSessionId,
-  ) => Effect.Effect<number, AuthControlPlaneError>;
-}
-
-export class AuthControlPlane extends Context.Service<AuthControlPlane, AuthControlPlaneShape>()(
-  "t3/auth/Services/AuthControlPlane",
-) {}
