--- apps/server/src/auth/Services/ServerAuth.ts
@@ -1,121 +1,0 @@
-import type {
-  AuthAccessTokenResult,
-  AuthBrowserSessionResult,
-  AuthClientMetadata,
-  AuthClientSession,
-  AuthCreatePairingCredentialInput,
-  AuthEnvironmentScope,
-  AuthPairingLink,
-  AuthPairingCredentialResult,
-  AuthSessionId,
-  AuthSessionState,
-  ServerAuthDescriptor,
-  ServerAuthSessionMethod,
-  AuthWebSocketTicketResult,
-} from "@t3tools/contracts";
-import * as Data from "effect/Data";
-import * as DateTime from "effect/DateTime";
-import * as Context from "effect/Context";
-import type * as Effect from "effect/Effect";
-import type * as HttpServerRequest from "effect/unstable/http/HttpServerRequest";
-
-export interface AuthenticatedSession {
-  readonly sessionId: AuthSessionId;
-  readonly subject: string;
-  readonly method: ServerAuthSessionMethod;
-  readonly scopes: ReadonlyArray<AuthEnvironmentScope>;
-  readonly expiresAt?: DateTime.DateTime;
-}
-
-export class ServerAuthInternalError extends Data.TaggedError("ServerAuthInternalError")<{
-  readonly message: string;
-  readonly cause?: unknown;
-}> {}
-
-export class ServerAuthInvalidCredentialError extends Data.TaggedError(
-  "ServerAuthInvalidCredentialError",
-)<{
-  readonly reason: "missing_credential" | "invalid_credential";
-  readonly cause?: unknown;
-}> {}
-
-export class ServerAuthInvalidRequestError extends Data.TaggedError(
-  "ServerAuthInvalidRequestError",
-)<{
-  readonly reason: "invalid_scope" | "scope_not_granted";
-}> {}
-
-export class ServerAuthForbiddenOperationError extends Data.TaggedError(
-  "ServerAuthForbiddenOperationError",
-)<{
-  readonly reason: "current_session_revoke_not_allowed";
-}> {}
-
-export interface ServerAuthShape {
-  readonly getDescriptor: () => Effect.Effect<ServerAuthDescriptor>;
-  readonly getSessionState: (
-    request: HttpServerRequest.HttpServerRequest,
-  ) => Effect.Effect<AuthSessionState, ServerAuthInternalError>;
-  readonly createBrowserSession: (
-    credential: string,
-    requestMetadata: AuthClientMetadata,
-  ) => Effect.Effect<
-    {
-      readonly response: AuthBrowserSessionResult;
-      readonly sessionToken: string;
-    },
-    ServerAuthInvalidCredentialError | ServerAuthInternalError
-  >;
-  readonly exchangeBootstrapCredentialForAccessToken: (
-    credential: string,
-    requestedScopes: ReadonlyArray<AuthEnvironmentScope> | undefined,
-    requestMetadata: AuthClientMetadata,
-  ) => Effect.Effect<
-    AuthAccessTokenResult,
-    ServerAuthInvalidCredentialError | ServerAuthInvalidRequestError | ServerAuthInternalError
-  >;
-  readonly issuePairingCredential: (
-    input?: AuthCreatePairingCredentialInput,
-  ) => Effect.Effect<AuthPairingCredentialResult, ServerAuthInternalError>;
-  readonly issueStartupPairingCredential: () => Effect.Effect<
-    AuthPairingCredentialResult,
-    ServerAuthInternalError
-  >;
-  readonly listPairingLinks: () => Effect.Effect<
-    ReadonlyArray<AuthPairingLink>,
-    ServerAuthInternalError
-  >;
-  readonly revokePairingLink: (id: string) => Effect.Effect<boolean, ServerAuthInternalError>;
-  readonly listClientSessions: (
-    currentSessionId: AuthSessionId,
-  ) => Effect.Effect<ReadonlyArray<AuthClientSession>, ServerAuthInternalError>;
-  readonly revokeClientSession: (
-    currentSessionId: AuthSessionId,
-    targetSessionId: AuthSessionId,
-  ) => Effect.Effect<boolean, ServerAuthForbiddenOperationError | ServerAuthInternalError>;
-  readonly revokeOtherClientSessions: (
-    currentSessionId: AuthSessionId,
-  ) => Effect.Effect<number, ServerAuthInternalError>;
-  readonly authenticateHttpRequest: (
-    request: HttpServerRequest.HttpServerRequest,
-  ) => Effect.Effect<
-    AuthenticatedSession,
-    ServerAuthInvalidCredentialError | ServerAuthInternalError
-  >;
-  readonly authenticateWebSocketUpgrade: (
-    request: HttpServerRequest.HttpServerRequest,
-  ) => Effect.Effect<
-    AuthenticatedSession,
-    ServerAuthInvalidCredentialError | ServerAuthInternalError
-  >;
-  readonly issueWebSocketTicket: (
-    session: Pick<AuthenticatedSession, "sessionId">,
-  ) => Effect.Effect<AuthWebSocketTicketResult, ServerAuthInternalError>;
-  readonly issueStartupPairingUrl: (
-    baseUrl: string,
-  ) => Effect.Effect<string, ServerAuthInternalError>;
-}
-
-export class ServerAuth extends Context.Service<ServerAuth, ServerAuthShape>()(
-  "t3/auth/Services/ServerAuth",
-) {}
