--- apps/server/src/auth/Services/SessionCredentialService.ts
@@ -1,103 +1,0 @@
-import type {
-  AuthClientMetadata,
-  AuthClientSession,
-  AuthEnvironmentScope,
-  AuthSessionId,
-  ServerAuthSessionMethod,
-} from "@t3tools/contracts";
-import * as Data from "effect/Data";
-import * as DateTime from "effect/DateTime";
-import * as Duration from "effect/Duration";
-import * as Context from "effect/Context";
-import type * as Effect from "effect/Effect";
-import type * as Stream from "effect/Stream";
-
-export interface IssuedSession {
-  readonly sessionId: AuthSessionId;
-  readonly token: string;
-  readonly method: ServerAuthSessionMethod;
-  readonly client: AuthClientMetadata;
-  readonly expiresAt: DateTime.DateTime;
-  readonly scopes: ReadonlyArray<AuthEnvironmentScope>;
-}
-
-export interface VerifiedSession {
-  readonly sessionId: AuthSessionId;
-  readonly token: string;
-  readonly method: ServerAuthSessionMethod;
-  readonly client: AuthClientMetadata;
-  readonly expiresAt?: DateTime.DateTime;
-  readonly subject: string;
-  readonly scopes: ReadonlyArray<AuthEnvironmentScope>;
-}
-
-export type SessionCredentialChange =
-  | {
-      readonly type: "clientUpserted";
-      readonly clientSession: AuthClientSession;
-    }
-  | {
-      readonly type: "clientRemoved";
-      readonly sessionId: AuthSessionId;
-    };
-
-export class SessionCredentialInvalidError extends Data.TaggedError(
-  "SessionCredentialInvalidError",
-)<{
-  readonly message: string;
-  readonly cause?: unknown;
-}> {}
-
-export class SessionCredentialInternalError extends Data.TaggedError(
-  "SessionCredentialInternalError",
-)<{
-  readonly message: string;
-  readonly cause?: unknown;
-}> {}
-
-export type SessionCredentialError = SessionCredentialInvalidError | SessionCredentialInternalError;
-
-export interface SessionCredentialServiceShape {
-  readonly cookieName: string;
-  readonly issue: (input?: {
-    readonly ttl?: Duration.Duration;
-    readonly subject?: string;
-    readonly method?: ServerAuthSessionMethod;
-    readonly scopes?: ReadonlyArray<AuthEnvironmentScope>;
-    readonly client?: AuthClientMetadata;
-  }) => Effect.Effect<IssuedSession, SessionCredentialInternalError>;
-  readonly verify: (token: string) => Effect.Effect<VerifiedSession, SessionCredentialError>;
-  readonly issueWebSocketToken: (
-    sessionId: AuthSessionId,
-    input?: {
-      readonly ttl?: Duration.Duration;
-    },
-  ) => Effect.Effect<
-    {
-      readonly token: string;
-      readonly expiresAt: DateTime.DateTime;
-    },
-    SessionCredentialInternalError
-  >;
-  readonly verifyWebSocketToken: (
-    token: string,
-  ) => Effect.Effect<VerifiedSession, SessionCredentialError>;
-  readonly listActive: () => Effect.Effect<
-    ReadonlyArray<AuthClientSession>,
-    SessionCredentialInternalError
-  >;
-  readonly streamChanges: Stream.Stream<SessionCredentialChange>;
-  readonly revoke: (
-    sessionId: AuthSessionId,
-  ) => Effect.Effect<boolean, SessionCredentialInternalError>;
-  readonly revokeAllExcept: (
-    sessionId: AuthSessionId,
-  ) => Effect.Effect<number, SessionCredentialInternalError>;
-  readonly markConnected: (sessionId: AuthSessionId) => Effect.Effect<void, never>;
-  readonly markDisconnected: (sessionId: AuthSessionId) => Effect.Effect<void, never>;
-}
-
-export class SessionCredentialService extends Context.Service<
-  SessionCredentialService,
-  SessionCredentialServiceShape
->()("t3/auth/Services/SessionCredentialService") {}
