+++ apps/server/src/cli/auth.ts
@@ -10,12 +10,8 @@ import * as Option from "effect/Option";
 import * as References from "effect/References";
 import { Argument, Command, Flag, GlobalFlag } from "effect/unstable/cli";
 
-import { AuthControlPlaneRuntimeLive } from "../auth/Layers/AuthControlPlane.ts";
-import {
-  AuthControlPlane,
-  INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT,
-} from "../auth/Services/AuthControlPlane.ts";
-import type { AuthControlPlaneShape } from "../auth/Services/AuthControlPlane.ts";
+import * as EnvironmentAuth from "../auth/EnvironmentAuth.ts";
+
 import {
   formatIssuedPairingCredential,
   formatIssuedSession,
@@ -30,9 +26,9 @@ import {
   resolveCliAuthConfig,
 } from "./config.ts";
 
-const runWithAuthControlPlane = <A, E>(
+const runWithEnvironmentAuth = <A, E>(
   flags: CliAuthLocationFlags,
-  run: (authControlPlane: AuthControlPlaneShape) => Effect.Effect<A, E>,
+  run: (environmentAuth: EnvironmentAuth.EnvironmentAuthShape) => Effect.Effect<A, E>,
   options?: {
     readonly quietLogs?: boolean;
   },
@@ -42,11 +38,11 @@ const runWithAuthControlPlane = <A, E>(
     const config = yield* resolveCliAuthConfig(flags, logLevel);
     const minimumLogLevel = options?.quietLogs ? "Error" : config.logLevel;
     return yield* Effect.gen(function* () {
-      const authControlPlane = yield* AuthControlPlane;
-      return yield* run(authControlPlane);
+      const environmentAuth = yield* EnvironmentAuth.EnvironmentAuth;
+      return yield* run(environmentAuth);
     }).pipe(
       Effect.provide(
-        Layer.mergeAll(AuthControlPlaneRuntimeLive).pipe(
+        Layer.mergeAll(EnvironmentAuth.runtimeLayer).pipe(
           Layer.provide(Layer.succeed(ServerConfig, config)),
           Layer.provide(Layer.succeed(References.MinimumLogLevel, minimumLogLevel)),
         ),
@@ -94,11 +90,11 @@ const pairingCreateCommand = Command.make("create", {
 }).pipe(
   Command.withDescription("Issue a new client pairing token."),
   Command.withHandler((flags) =>
-    runWithAuthControlPlane(
+    runWithEnvironmentAuth(
       flags,
-      (authControlPlane) =>
+      (environmentAuth) =>
         Effect.gen(function* () {
-          const issued = yield* authControlPlane.createPairingLink({
+          const issued = yield* environmentAuth.createPairingLink({
             scopes: AuthStandardClientScopes,
             subject: "one-time-token",
             ...(Option.isSome(flags.ttl) ? { ttl: flags.ttl.value } : {}),
@@ -123,12 +119,12 @@ const pairingListCommand = Command.make("list", {
 }).pipe(
   Command.withDescription("List active client pairing tokens without revealing their secrets."),
   Command.withHandler((flags) =>
-    runWithAuthControlPlane(
+    runWithEnvironmentAuth(
       flags,
-      (authControlPlane) =>
+      (environmentAuth) =>
         Effect.gen(function* () {
-          const pairingLinks = yield* authControlPlane.listPairingLinks({
-            excludeSubjects: [INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT],
+          const pairingLinks = yield* environmentAuth.listPairingLinks({
+            excludeSubjects: [EnvironmentAuth.INTERNAL_ADMINISTRATIVE_BOOTSTRAP_SUBJECT],
           });
           yield* Console.log(formatPairingCredentialList(pairingLinks, { json: flags.json }));
         }),
@@ -145,9 +141,9 @@ const pairingRevokeCommand = Command.make("revoke", {
 }).pipe(
   Command.withDescription("Revoke an active client pairing token."),
   Command.withHandler((flags) =>
-    runWithAuthControlPlane(flags, (authControlPlane) =>
+    runWithEnvironmentAuth(flags, (environmentAuth) =>
       Effect.gen(function* () {
-        const revoked = yield* authControlPlane.revokePairingLink(flags.id);
+        const revoked = yield* environmentAuth.revokePairingLink(flags.id);
         yield* Console.log(
           revoked
             ? `Revoked pairing credential ${flags.id}.\n`
@@ -173,11 +169,11 @@ const sessionIssueCommand = Command.make("issue", {
 }).pipe(
   Command.withDescription("Issue a scoped bearer access token for headless or remote clients."),
   Command.withHandler((flags) =>
-    runWithAuthControlPlane(
+    runWithEnvironmentAuth(
       flags,
-      (authControlPlane) =>
+      (environmentAuth) =>
         Effect.gen(function* () {
-          const issued = yield* authControlPlane.issueSession({
+          const issued = yield* environmentAuth.issueSession({
             scopes: AuthAdministrativeScopes,
             ...(Option.isSome(flags.ttl) ? { ttl: flags.ttl.value } : {}),
             ...(Option.isSome(flags.label) ? { label: flags.label.value } : {}),
@@ -203,11 +199,11 @@ const sessionListCommand = Command.make("list", {
 }).pipe(
   Command.withDescription("List active sessions without revealing bearer tokens."),
   Command.withHandler((flags) =>
-    runWithAuthControlPlane(
+    runWithEnvironmentAuth(
       flags,
-      (authControlPlane) =>
+      (environmentAuth) =>
         Effect.gen(function* () {
-          const sessions = yield* authControlPlane.listSessions();
+          const sessions = yield* environmentAuth.listSessions();
           yield* Console.log(formatSessionList(sessions, { json: flags.json }));
         }),
       {
@@ -226,9 +222,9 @@ const sessionRevokeCommand = Command.make("revoke", {
 }).pipe(
   Command.withDescription("Revoke an active session."),
   Command.withHandler((flags) =>
-    runWithAuthControlPlane(flags, (authControlPlane) =>
+    runWithEnvironmentAuth(flags, (environmentAuth) =>
       Effect.gen(function* () {
-        const revoked = yield* authControlPlane.revokeSession(flags.sessionId);
+        const revoked = yield* environmentAuth.revokeSession(flags.sessionId);
         yield* Console.log(
           revoked
             ? `Revoked session ${flags.sessionId}.\n`
