+++ apps/server/src/cliAuthFormat.ts
@@ -1,7 +1,7 @@
 import type { AuthClientMetadata, AuthClientSession, AuthPairingLink } from "@t3tools/contracts";
 import * as DateTime from "effect/DateTime";
 
-import type { IssuedBearerSession, IssuedPairingLink } from "./auth/Services/AuthControlPlane.ts";
+import type { IssuedBearerSession, IssuedPairingLink } from "./auth/EnvironmentAuth.ts";
 
 const newline = "\n";
